+++ project/spark/hooks/post_gen_project.py
@@ -9,7 +9,7 @@ import logging
 MANIFEST = "manifest.yml"
 GROUP_ID = "{{ cookiecutter.group_id }}"
 MODULE_NAME = "{{ cookiecutter.module_name }}"
-datafy_managed_role = "{{ cookiecutter.datafy_managed_role }}"
+conveyor_managed_role = "{{ cookiecutter.conveyor_managed_role }}"
 cloud = "{{ cookiecutter.cloud }}"
 project_type = "{{ cookiecutter.project_type }}"
 
@@ -38,7 +38,7 @@ def create_group_id_directories():
 
 
 def cleanup_resources():
-    if not bool(strtobool(datafy_managed_role)) or cloud == "azure":
+    if not bool(strtobool(conveyor_managed_role)) or cloud == "azure":
         shutil.rmtree("resources")