+++ project/spark/{{ cookiecutter.project_name }}/dags/{{ cookiecutter.project_name }}.py
@@ -1,10 +1,10 @@
 from airflow import DAG
-from datafy.operators import DatafySparkSubmitOperatorV2
+from conveyor.operators import ConveyorSparkSubmitOperatorV2
 from airflow.utils import dates
 from datetime import timedelta
 
 default_args = {
-    "owner": "Datafy",
+    "owner": "Conveyor",
     "depends_on_past": False,
     "start_date": dates.days_ago(2),
     "email": [],
@@ -22,7 +22,7 @@ dag = DAG(
     max_active_runs=1,
 )
 
-sample_task = DatafySparkSubmitOperatorV2(
+sample_task = ConveyorSparkSubmitOperatorV2(
     dag=dag,
     task_id="sample",
     num_executors="1",
@@ -33,14 +33,14 @@ sample_task = DatafySparkSubmitOperatorV2(
     {%- elif cookiecutter.spark_version == "3.0" -%}
     spark_main_version=3,
     {%- endif %}
-    { % - if cookiecutter.datafy_managed_role %}
-    aws_role="{{ cookiecutter.project_name }}-{% raw %}{{ macros.datafy.env() }}{% endraw %}",
-    { % - endif %}
+    {% if cookiecutter.conveyor_managed_role -%}
+    aws_role="{{ cookiecutter.project_name }}-{% raw %}{{ macros.conveyor.env() }}{% endraw %}",
+    {%- endif %}
     {% if cookiecutter.spark_version == "2.4" -%}
     application="/opt/spark/work-dir/app.jar",
     {%- elif cookiecutter.spark_version == "3.0" -%}
     application="local:///opt/spark/work-dir/app.jar",
     {%- endif %}
     java_class="{{ cookiecutter.group_id }}.{{ cookiecutter.module_name }}.SampleJob",
-    application_args=[{% raw %}"--date", "{{ ds }}", "--environment", "{{ macros.datafy.env() }}"{% endraw %}],
+    application_args=[{% raw %}"--date", "{{ ds }}", "--environment", "{{ macros.conveyor.env() }}"{% endraw %}],
 )