+++ tests/projects/test_python.py
@@ -20,7 +20,7 @@ def test_python_template(cookies):
 def test_python_template_no_role(cookies):
     result = cookies.bake(
         template=f"{os.path.dirname(os.path.abspath(__file__))}/../../project/python",
-        extra_context={"datafy_managed_role": False},
+        extra_context={"conveyor_managed_role": False},
     )
     assert 0 == result.exit_code
     assert result.exception is None